refactor: 컬렉션뷰 헤더의 이름과 전체보기 버튼을 MainCollectionTitleHeaderView로 분리
